chore: reconfigure runtime for docker support
