+++ app/__main__.py
@@ -6,4 +6,3 @@ app = create_app()
 if __name__ == "__main__":
     logging.info("Flask app started.")
     app.run(host="0.0.0.0", port=8000)
-
